backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & Migrations
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# Environment & Configuration
python-dotenv==1.0.0

# Trading Platform Integrations
MetaTrader5==5.0.45
python-binance==1.0.19
ccxt==4.2.25
nsepy==0.8
yfinance==0.2.35

# Machine Learning & AI
torch==2.2.0
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.2
prophet==1.1.5
statsmodels==0.14.1
tensorflow==2.15.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
pandas-ta==0.3.14b
ta==0.11.0
pyti==0.2.2

# Background Tasks
celery==5.3.6

# Async & WebSocket
websockets==12.0
python-socketio==5.11.0
aiohttp==3.9.3
httpx==0.26.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.2

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.40.0

# Utilities
tenacity==8.2.3
loguru==0.7.2
pyyaml==6.0.1

